+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 All notable changes to FP Updater will be documented in this file.
 
 ## [1.6.10] - 2026-03-15
-### Fixed
-- Eliminazione cliente: i clienti rimossi dalla lista non riappaiono più quando il sito cliente si riconnette al Master (blacklist clienti rimossi)
+### Added
+- Alias client_id dopo rinomina: se rinomini un cliente, il sito che si riconnette con il vecchio ID aggiorna l'entry con il nome nuovo (il vecchio nome non riappare)
 
 ## [1.6.9] - 2026-03-10
 ### Changed